+++ Source/Forms/Gen3/Wild3.cpp
@@ -153,7 +153,7 @@ void Wild3::updateLocationsGenerator()
     auto encounter = static_cast<Encounter>(ui->comboBoxGeneratorEncounter->currentData().toInt());
     encounterGenerator = Encounters3::getEncounters(encounter, currentProfile.getVersion());
 
-    std::vector<u8> locs;
+    std::vector<u16> locs;
     std::transform(encounterGenerator.begin(), encounterGenerator.end(), std::back_inserter(locs),
                    [](const EncounterArea3 &area) { return area.getLocation(); });
 
@@ -174,7 +174,7 @@ void Wild3::updateLocationsSearcher()
     auto encounter = static_cast<Encounter>(ui->comboBoxSearcherEncounter->currentData().toInt());
     encounterSearcher = Encounters3::getEncounters(encounter, currentProfile.getVersion());
 
-    std::vector<u8> locs;
+    std::vector<u16> locs;
     std::transform(encounterSearcher.begin(), encounterSearcher.end(), std::back_inserter(locs),
                    [](const EncounterArea3 &area) { return area.getLocation(); });
